core: widen the request message handler lambdas
MavlinkRequestMessageHandler::Callback passes a 32 bit system ID, but
these three lambdas still took it as uint8_t. On 64 bit that narrowing
is C4244, which the Windows jobs don't enable, so it went unnoticed. On
32 bit Windows size_t is unsigned int and therefore the same type as
uint32_t, which makes MSVC report it as C4267 instead, and that one is
on and fatal under WERROR.

+++ cpp/src/mavsdk/core/server_component_impl.cpp
@@ -49,7 +49,7 @@ ServerComponentImpl::ServerComponentImpl(
 
     _mavlink_request_message_handler.register_handler(
         MAVLINK_MSG_ID_PROTOCOL_VERSION,
-        [this](uint8_t, uint8_t, const MavlinkRequestMessageHandler::Params&) {
+        [this](uint32_t, uint8_t, const MavlinkRequestMessageHandler::Params&) {
             send_protocol_version();
             return MAV_RESULT_ACCEPTED;
         },
@@ -57,7 +57,7 @@ ServerComponentImpl::ServerComponentImpl(
 
     _mavlink_request_message_handler.register_handler(
         MAVLINK_MSG_ID_AUTOPILOT_VERSION,
-        [this](uint8_t, uint8_t, const MavlinkRequestMessageHandler::Params&) {
+        [this](uint32_t, uint8_t, const MavlinkRequestMessageHandler::Params&) {
             send_autopilot_version();
             return MAV_RESULT_ACCEPTED;
         },
